refactor(cli): extract project batch into commands/project.ts (5.0-alpha)
Fourteenth cut of the cli.ts god-module split — the project-lifecycle leaves.

- New src/commands/project.ts — cmdCreatePlugin, cmdClone, cmdRun, cmdOpen
  (all exported). cmdClone runs cmdSetup in the cloned repo, imported from the
  sibling setup module (no cycle: setup does not import project).
- cli.ts imports them back; COMMANDS table unchanged. Pruned now-unused imports
  (createPlugin, cloneRepository, executeCommand, listServices/openService).

Verification: tsc clean, eslint 0 errors, full suite 2580 passing, prettier
clean, runtime dispatch smoke-tested (create-plugin/run/open).

cli.ts: 1859 -> 1644 lines (-215; -4686 cumulative from the 6330-line original,
~74% smaller).
